fix bug according to SonarQube
This is just to test it out.

+++ src/main/java/pr_se/gogame/model/ruleset/JapaneseRuleset.java
@@ -82,7 +82,7 @@ public class JapaneseRuleset implements Ruleset {
         }
 
         double komi = game.getKomi();
-        int handicap = game.getHandicap();
+        double handicap = game.getHandicap();
 
         int capturedStonesBlack = game.getStonesCapturedBy(StoneColor.BLACK);
         int territoryScoreBlack = calculateTerritoryPoints(StoneColor.BLACK, game);
